VRSL 2.4.1 Update
- Added Alternative Strobe texture settings. Toggle to switch from traditional dynamic DMX strobing to a 3 speed strobe channel. Affects all fixtures.
- Adjusted all VRSL materials and a few textures to prepare for GI release.
- Added ability to edit VRSL DMX Custom Render Textures in the VRSL Control Panel.
- Added ability to set 2D Noise Texture settings between high and medium modes for the spotlight and washlight volumetric DMX shaders.

+++ Packages/com.acchosen.vr-stage-lighting/Runtime/Shaders/Discoball/VRSL-Discoball.shader
@@ -151,7 +151,7 @@
                         4.0 / 17.0, 12.0 / 17.0,  2.0 / 17.0, 10.0 / 17.0,
                         16.0 / 17.0,  8.0 / 17.0, 14.0 / 17.0,  6.0 / 17.0
                     };
-                    int index = (int(pos.x) % 4) * 4 + int(pos.y) % 4;
+                    int index = (int)((uint(pos.x) % 4) * 4 + uint(pos.y) % 4);
 		        #endif
                 float4 depthdirect = i.worldDirection * (1.0f / i.vertex.w);
                 float sceneZ = SAMPLE_DEPTH_TEXTURE(_CameraDepthTexture, i.screenPos.xy / i.screenPos.w);
